Removed the brk variable and the code for setting it in the processor status.
From: http://visual6502.org/wiki/index.php?title=6502_BRK_and_B_bit

the B flag position in the status register is not a bit in the status
register: it is unaffected by PLP and RTI. However, the 6502 does push
the register with either a 1 or 0. The intention is to distinguish a BRK
from an IRQ, which is needed because these two share the same vector.